Style: model/backend split
